integrate auth endpoints | Auth React Context

+++ client/src/api/auth.ts
@@ -1,8 +1,15 @@
+import { IUser } from "../App";
 import api from "./api";
 
-export const login = (body: any, config = {}) =>
-  api.post("login", body, config).then((res) => res.data);
-export const register = (body: any, config = {}) =>
-  api.post(`register`, body, config).then((res) => res.data); // TODO body interface for register and login
+export interface IAuthForm {
+  name?: string;
+  username: string;
+  password: string;
+}
+
+export const login = (body: IAuthForm, config = {}): Promise<IUser> =>
+  api.post("auth/login", body, config).then((res) => res.data);
+export const register = (body: IAuthForm, config = {}): Promise<IUser> =>
+  api.post(`auth/register`, body, config).then((res) => res.data); // TODO body interface for register and login
 export const logout = (userId: string, config = {}) =>
-  api.delete(`logout/${userId}`, config).then((res) => res.data);
+  api.delete(`auth/logout/${userId}`, config).then((res) => res.data);
